add tariff .xml .xsd

+++ src/main/java/com/epam/eugene/entity/CallPrices.java
@@ -1,37 +1,37 @@
 package com.epam.eugene.entity;
 
 public class CallPrices {
-    private double priceWithinTheNetwork;
-    private double priceOutsideNetwork;
-    private double pricelandlinePhones;
+    private float priceWithinTheNetwork;
+    private float priceOutsideNetwork;
+    private float pricelandlinePhones;
 
-    public CallPrices(double priceWithinTheNetwork, double priceOutsideNetwork, double pricelandlinePhones) {
+    public CallPrices(float priceWithinTheNetwork, float priceOutsideNetwork, float pricelandlinePhones) {
         this.priceWithinTheNetwork = priceWithinTheNetwork;
         this.priceOutsideNetwork = priceOutsideNetwork;
         this.pricelandlinePhones = pricelandlinePhones;
     }
 
-    public double getPriceWithinTheNetwork() {
+    public float getPriceWithinTheNetwork() {
         return priceWithinTheNetwork;
     }
 
-    public void setPriceWithinTheNetwork(double priceWithinTheNetwork) {
+    public void setPriceWithinTheNetwork(float priceWithinTheNetwork) {
         this.priceWithinTheNetwork = priceWithinTheNetwork;
     }
 
-    public double getPriceOutsideNetwork() {
+    public float getPriceOutsideNetwork() {
         return priceOutsideNetwork;
     }
 
-    public void setPriceOutsideNetwork(double priceOutsideNetwork) {
+    public void setPriceOutsideNetwork(float priceOutsideNetwork) {
         this.priceOutsideNetwork = priceOutsideNetwork;
     }
 
-    public double getPricelandlinePhones() {
+    public float getPricelandlinePhones() {
         return pricelandlinePhones;
     }
 
-    public void setPricelandlinePhones(double pricelandlinePhones) {
+    public void setPricelandlinePhones(float pricelandlinePhones) {
         this.pricelandlinePhones = pricelandlinePhones;
     }
